+++ internal/app/room/usecase.go
@@ -4,7 +4,7 @@ import roomModel "github.com/Kostikans/avitoTest/internal/app/room/model"
 
 type Usecase interface {
 	AddRoom(room roomModel.RoomAdd) (roomModel.RoomID, error)
-	DeleteRoom(roomID int64) error
+	DeleteRoom(roomID int) error
 	GetRooms(order roomModel.RoomOrder) ([]roomModel.Room, error)
-	CheckRoomExist(roomID int64) (bool, error)
+	CheckRoomExist(roomID int) (bool, error)
 }
